Loading a project only fails when really necessary.
As long as only parts of the project file could not be read the remainder is still loaded and can be used. This allows to work with partially corrupted project files.

Only when no audio device to open was found a fatal error is raised and the project will not be loaded. If another device than the one configured in the project could be opened only a warning is produced and not a fatal error.