Support file drag'n'drop
Audio files can be dropped onto a track. The files are put onto that track and the tracks below.
Audio files and image files can be dropped onto a Jingle Player. The audio files are put onto the tracks of the player, beginning with the first track. The first image file is used as the image for the player.

closes #3