Fix hidden time display in players
The text became invisible after loading a project. This is because by default players have a transparent background color. When this is loaded an inverse color is selected for the text which is also transparent. Therefore the text was hidden.

The solution is to check for a transparent background and resort to black.