Support all playback sample rates in CD Player
The sample positions of the audio CD and the transport source were used interchangeably. This is incorrect because audio CDs are always 44.1 kHz while the transport source has the sample rate of the playback device.

The fix is to convert sample positions from the audio CD before passing it to the transport source and directly reading the sample position from the source instead of the transport source.