Pass parent XmlElement to all saveToXml methods
When loading an XML file the caller of restoreFromXml must know the element name. When saving an XML file the element name was sometimes set by the saveToXml method. This means the element name must be synchronized across classes. To only have a single class the element is created by the caller and passed to saveToXml.