Fix crash when saving in tabs view
Only the active window has a parent component which is why saving
crashed when there was more than one tab. The bounds of the window are
stored in the mdiDocumentPos_ property when a floating window becomes a
tab. Therefore this is stored when the view is set to tabs. As before,
the bounds of the parent (which is the window of the player) are stored
when the view is set to windows.

When loading a project either the Bounds or the MdiDocumentPos element
must exist for each player and is loaded to give the player its bounds
from when the project was saved.