Set audio thumbnail thread to lowest priority
By default AudioThumbnailCache starts its thread with a priority of 2.
This is translated to THREAD_PRIORITY_BELOW_NORMAL on Windows
resulting in a base priority of 7. Compared to the default priority
of 8 this is farily high. See [1] for details.

To make sure the player has enough ressources for playback the
priority of the thumbnail thread is reduced to to the lowest
possible value.

[1]: https://msdn.microsoft.com/de-de/library/windows/desktop/ms685100