Support sending MTC during playback
The behavior of jumping to a position while the not playing was changed!

By clicking into a track while it is not playing it's possible to
jump to a position. When play is pressed the playback starts at that
position. The decision whether play starts at the position or from
the beginning is done inside the track. MTCSender is not informed
about the decision.

A simple solution is to always play from the beginning and use pause
to play from a position.

Disabled warning C4458 because this is not an issue in local scope.