Drastically improve performance of volume calculation
Until now the volume was calculated by keeping a circular buffer 1/10th
of the sample rate (i.e. 100 ms) and determining the absolute maximum
sample within it. This is immensly CPU-intensive because the maximum of
quite a lot of numbers needs to be recalculated very often.

The new approach was taken from the JUCE forum[1] which is much easier
to calculate. It will not be as "perfect" as the previous solution but
should be good enough.

[1] https://forum.juce.com/t/level-meter-feedback/31161/12