Request zero input channels
Until now MStarPlayer asked for up to 64 input channels when the
AudioDeviceManager was initialized. At least on Linux with ALSA this
lead to significant performance impact as the input data was processed
by code in the JUCE library. MStarPlayer doesn't support audio input so
this was completely unnecessary.

By requesting zero input channels the CPU usage can drop by up to 50 %.