Fix handling of auto-playing next playlist entry
If a playlist entry is marked as having to play the following entry this
will be done so in the position callback. This callback was processed
before the special case of the playback having stopped. This meant that
if a playlist entry finished it would first start playback of the next
file and then behave as if the playback reached its end.

This lead to subtle bugs like MTC continuing to progress or the UI not
updating correctly.

Fixed it by first handling the case that the playback ended and then
call the position callbacks which potentially start a new playback.