Ensure correct player color is shown when stopped
There was a bug report that a JinglePlayer showed the alternate color
used during the blink animation that happens in the last 10 seconds of
playback.

The same situation arises reproducibly when the audio file is shorter
than 10 seconds and the millisecond part is greater than 500 ms. In that
case the alternate color for blinking is always shown when the playback
is stopped.

The fix for both is to reset `m_blink` at the end of the playback.