Fix failing build
The previous commit changed the MTCSender but was not based on the
latest master commit. The master branch contained additional code to
load and store the currently selected MIDI device with the project.

This lead to a build break because the API of MTCSender changed. The fix
is to adjust the code to store and load the selected devices with the
project. Loading old projects still works by checking for the old way
the single selected device was stored.