Update to JUCE 6.1.0
All code changes are related to this breaking change in JUCE:

> The default value of JUCE_MODAL_LOOPS_PERMITTED has been changed from 1 to 0.

The changes can be grouped in three parts:

asynchronous FileChoosers
-------------------------

All usage of the FileChooser became asynchronous.

The file dialog usage in `askSaveProject` required some additional
refactoring. With the `requestX`-type functions unsaved modifications
will not get lost. The user will be prompted to first save his changes.
With the `uncheckedX`-type functions current modifications will be
dropped silently. The usual flow is:

    requestX -> askForSaveProject -> uncheckedX

The `m_currentFileChooser` members ensure two things:

1. A file dialog is closed when it becomes obsolete, e.g. when the
   player was closed.
2. Each context only keeps open a single dialog. This will for example
   prevent two open dialogs at the same time.

Popup menus
-----------

Those were shown synchronously with a following switch statement that
performed the correct action for the selected menu item. To be
asynchronous the code for each item is now in a callback which is given
to `addItem`. This keeps the item and its code closer and allows showing
the popup asynchronously.

**Potential issue:** The popups have no parent item so they might live
longer than the bound `this`. For that to be destroyed some user
interaction is required but each user interaction closes the popup, so
this should not be a problem.

CrashDumper
-----------

While a crash dump is written on Windows a dialog is shown. This dialog
can no longer block the exception filter so a future was introduced that
gets resolved once the dump file was written.