Update to JUCE 8.0.1
In JUCE 7.0.3 `Thread::setPriority (int)` was removed[^1]. The thread
for reading audio files ahead uses a low priority now which roughly
matches the previous numeric value. The audio thumbnail cache thread is
initialized by JUCE with a low priority. This priority can't be changed
from the outside so it should be fine.

[^1]: https://github.com/juce-framework/JUCE/blob/8.0.1/BREAKING_CHANGES.md#change-49