Allow quick (de)activation of output channels
For audio devices with a high number of output channels it can be
tedious having to activate or deactivate most channels. Offer buttons
to do either very quickly so users only need to activate or deactivate a
few more channels.

The first channel won't be deactivated since JUCE requires at least one
channel to be active.