Fix crash with empty track
When a track hasn't loaded a file the `m_transportSource` won't contain
a source and thus crashes on some interactions (but not on others).
Protect the calls by first checking whether an audio file was loaded.

Also call the *playing state changed* callback since the caller assumes
the change actually changed so it needs to be "reset" for the UI to
match the actual behavior.