Support panning stereo player tracks
Some code for this already existed but was always disabled because the
`panEnabled` parameter was always set to `false`. This commit now adds
the remaining changes necessary to be able to pan stereo player tracks.

The gain is applied in `ChannelRemappingAudioSourceWithVolume` similar
to how the output channel mapping works with an array indexed by the
source channel number.

The fader is implemented in `PanSlider` as a copy of the `VolumeSlider`
with different visuals but the same features. Specifically it also snaps
to a neutral position which is marked with a line.

The existing code placed the slider below the solo and mute buttons,
squeezing the volume slider but leaving an empty label above. This isn't
good use of the available space. Instead the pan slider now replaces the
label for tracks.

The pan slider works in the value range -1.0 to 1.0 from which the gain
values for both channels are calculated in `Track:setPan`.